% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{cross.tab}
\alias{cross.tab}
\title{cross.tab}
\usage{
cross.tab(dep, indep, cells = "prop.r", chisq = T, weight = NULL,
  digits = 3)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{weighs}{a numeric vector of weights}

\item{either}{"row", "column" or "count"}
}
\value{
a data.frame
}
\description{
Creates a frequency table from the column vectors in a data.frame.
}

