% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eyeball.R
\name{eyeball}
\alias{eyeball}
\title{eyeball}
\usage{
eyeball(dep, vars, weight = NULL, cell = "prop.c", p.value = 0.05)
}
\arguments{
\item{dep}{a dataframe of dependent variables}

\item{vars}{a dataframe of independent variables}

\item{weight}{}

\item{cell}{cell.proportions (only supports one form at the moment)}

\item{p.value}{}
}
\value{
an xlsx file with all tables with significant levels above p.value
}
\description{
multiple crosstables & tabout
}

