% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{freq.scale}
\alias{freq.scale}
\title{freq.scale}
\usage{
freq.scale(x, cells = c("count", "pct"), weight = NULL, header = NULL)
}
\arguments{
\item{x}{a data.frame}

\item{cells}{a vector indicating which cells to export.Default is c("count", "pct")}

\item{header}{add a header (for tabout)}

\item{weighs}{a numeric vector of weights}
}
\value{
a data.frame
}
\description{
Creates a frequency table from the column vectors in a data.frame.
}

