% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{freq.tab}
\alias{freq.tab}
\title{freq.tab}
\usage{
freq.tab(x, weights = NULL, transpose = FALSE)
}
\arguments{
\item{x}{a vector}

\item{transpose}{transposes the table if set as TRUE}

\item{weighs}{a numeric vector of weights}
}
\value{
a data.frame
}
\description{
Creates a simple frequency table and store it in a data.frame.
In additiion it adds attributes (header, sub.header, answer, type) to the output data.frame.
Attributes are usefull when writing tables (data.frames) to excel (with the tabout function)
}

