% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{report.freq}
\alias{report.freq}
\title{Create a report with frequency tables}
\usage{
report.freq(variables)
}
\arguments{
\item{variables}{is a data.frame of factors}
}
\value{
Several plotted tables
}
\description{
Takes a data.frame of factors and returns plotted frequency tables created by \link{tab.freq}
}

