% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simple_table.R
\name{standard.width.table}
\alias{standard.width.table}
\title{Standard width table}
\usage{
standard.width.table(tab, standard = 120)
}
\arguments{
\item{tab}{is a table}

\item{standard}{is the width of the table in number of characters}
}
\value{
a reformatted tabel
}
\description{
Formats a table to a standard width
}

