% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simple_table.R
\name{tab.freq}
\alias{tab.freq}
\title{Frequency-table}
\usage{
tab.freq(x, digits = 2)
}
\arguments{
\item{x}{is a factor or character vector}

\item{digits}{integer indicating the number of decimal places}
}
\value{
a table of frequencies and proportions
}
\description{
Frequency-table
}

