% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabout.R
\name{tabout}
\alias{tabout}
\title{tabout}
\usage{
tabout(x, sheet.name = NULL, overwrite = TRUE, file = "My_tables.xlsx",
  row.break.length = 2, write = TRUE, style = 1, dropif = 30,
<<<<<<< HEAD
  headers = names(x), header.pos = 1)
=======
<<<<<<< HEAD
  headers = names(x), header.pos = 1)
=======
  headers = names(x))
>>>>>>> e9f3c3f35edb28fcc70b9973b91c43d2058e16bb
>>>>>>> 5585abab7a5b37fa64ad4cd4be1ef9460d1adbe8
>>>>>>> f16214ecd4e6001ac36a5b8c62865093d1d5020d
}
\arguments{
\item{x}{a list or a data.frame}

\item{sheet.name}{name your sheet}

\item{overwrite}{}

\item{file}{filename}

\item{row.break.length}{number of empty rows between dataframes in sheet}

\item{write}{if false returns workbook}

\item{style}{chose how to style your frames (currently you can only  style=1)}
}
\description{
takes a list of dataframes and write them to a sheet
}

