% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tal_table.R
\name{tal.freq}
\alias{tal.freq}
\title{Frequency table in the &tal style}
\usage{
tal.freq(x, digits = 2, color = rgb(144, 26, 30, maxColorValue = 255),
  title = "", min.width = 120, footnote = "")
}
\arguments{
\item{x}{is a factor or character vector}

\item{digits}{integer indicating the number of decimal places}

\item{title}{sets the title}

\item{min.width}{is the minimum width of the table, see \link{standard.width.table}}

\item{footnote}{is the text displayed below the table}
}
\description{
Frequency table in the &tal style
}

