% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{freq.mc}
\alias{freq.mc}
\title{freq.mc}
\usage{
freq.mc(x, cells = c("count", "pct"), yes.answer = NULL, row.number = 1,
  weight = NULL)
}
\arguments{
\item{x}{a data.frame}

\item{yes.answer}{a character value that can be used to match which element of a column vector to use in creating
the multiple choice frequency table.}

\item{row.number}{a numerical value that can be used to match which element of a column vector to use in creating
the multiple choice frequency table.}
}
\value{
a data.frame
}
\description{
Creates a multiple choice frequency table from the column vectors in a data.frame.
}
\details{
When working with survey data one often finds oneself with a single variable for each element of a multiple choice question.
This function performs a frequency table on each variable in x and compounds the yes.answer into a single frequency table
and stores it in a data.frame.

If a column vector within x does not have an answer attribute acribed to it, then yes.answer and row.number can be used
to match which element of the vector the function should use instead.
}

