% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frequency_tables.R
\name{freq.tab}
\alias{freq.tab}
\title{freq.tab}
\usage{
freq.tab(x, transpose = FALSE, cells = c("count", "pct"), weights = NULL,
  digits = 3)
}
\arguments{
\item{x}{a vector}

\item{transpose}{transposes the table if set as TRUE}

\item{cells}{a vector indicating which cells to export.Default is c("count", "pct")}

\item{weighs}{a numeric vector of weights}
}
\value{
a data.frame
}
\description{
Creates a simple frequency table and stores it in a data.frame.
In additiion it adds attributes (header, sub.header, answer, type) to the output data.frame.
}

