% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_lime_data.R
\name{prepare.lime.data}
\alias{prepare.lime.data}
\title{prepare.lime.data}
\usage{
prepare.lime.data(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a data.frame
}
\description{
Adds a number of attributes (header, sub.header, answer) to each variable in a data.frame.
In addition it revalues the level "Yes" in multiple choice variables from Limesurvey.
}

